rename files to follow python naming
